refactor: Move CLI bootstrap and build info to internal packages
The command-line interface (CLI) setup logic, including flag
definitions, command structure, and the main execution flow, was moved
from the top-level `cmd/lazyworktree` directory into a new
`internal/bootstrap` package. This consolidates the CLI bootstrapping
concerns.

Additionally, a new `internal/buildinfo` package was created to manage
build metadata like version, commit hash, and build date. This
centralizes the handling of this information and allows for dynamic
enrichment from build environment details.

The `cmd/lazyworktree/main.go` file was updated to use the new
`internal/bootstrap` package for running the application and
`internal/buildinfo` to set build information. The existing test files
and command/flag definitions were also renamed and moved to reflect
their new locations within the `internal` directory structure.
