+++ internal/bootstrap/commands_exec_test.go
@@ -1,4 +1,4 @@
-package main
+package bootstrap
 
 import (
 	"os"
